requirements.txt
python-telegram-bot
google-generativeai
pillow
gspread
oauth2client
fastapi
uvicorn
pandas
openpyxl
python-jose[cryptography]
passlib[bcrypt]
python-multipart
